@@ -1,8 +1,8 @@
 package org.flyve.mdm.agent.core.deeplink;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.rule.ActivityTestRule;
+import androidx.test.InstrumentationRegistry;
+import androidx.test.rule.ActivityTestRule;
 
 import org.flyve.mdm.agent.ui.StartEnrollmentActivity;
 import org.junit.Assert;
