@@ -33,7 +33,7 @@ import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v4.content.FileProvider;
+import androidx.core.content.FileProvider;
 
 import org.flyve.mdm.agent.BuildConfig;
 import org.flyve.mdm.agent.R;
