@@ -23,10 +23,10 @@
 
 package org.flyve.mdm.agent.data.database.dao;
 
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.Query;
-import android.arch.persistence.room.Update;
+import androidx.room.Dao;
+import androidx.room.Insert;
+import androidx.room.Query;
+import androidx.room.Update;
 
 import org.flyve.mdm.agent.data.database.entity.MDMLog;
 
