@@ -27,9 +27,9 @@ package org.flyve.mdm.agent.data.database.entity;
  * ------------------------------------------------------------------------------
  */
 
-import android.arch.persistence.room.ColumnInfo;
-import android.arch.persistence.room.Entity;
-import android.arch.persistence.room.PrimaryKey;
+import androidx.room.ColumnInfo;
+import androidx.room.Entity;
+import androidx.room.PrimaryKey;
 
 @Entity(tableName = "topics")
 public class Topics {
