@@ -23,9 +23,9 @@
 
 package org.flyve.mdm.agent.data.database.setup;
 
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.Room;
-import android.arch.persistence.room.RoomDatabase;
+import androidx.room.Database;
+import androidx.room.Room;
+import androidx.room.RoomDatabase;
 import android.content.Context;
 
 import org.flyve.mdm.agent.data.database.dao.ApplicationDao;
