@@ -24,13 +24,15 @@
 package org.flyve.mdm.agent.ui;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.text.Html;
 import android.text.method.LinkMovementMethod;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.fragment.app.Fragment;
+
 import org.flyve.mdm.agent.R;
 import org.flyve.mdm.agent.utils.EnvInfoAbout;
 
