@@ -25,8 +25,6 @@ package org.flyve.mdm.agent.ui;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.widget.SwipeRefreshLayout;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -34,6 +32,10 @@ import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.ProgressBar;
 import android.widget.TextView;
+
+import androidx.fragment.app.Fragment;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
+
 import org.flyve.mdm.agent.R;
 import org.flyve.mdm.agent.adapter.ApplicationsAdapter;
 import org.flyve.mdm.agent.data.database.ApplicationData;
