@@ -25,7 +25,6 @@ package org.flyve.mdm.agent.ui;
 
 import android.os.Bundle;
 import android.os.Environment;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -33,6 +32,8 @@ import android.widget.Button;
 import android.widget.CompoundButton;
 import android.widget.EditText;
 import android.widget.Switch;
+
+import androidx.fragment.app.Fragment;
 
 import org.flyve.mdm.agent.R;
 import org.flyve.mdm.agent.data.database.PoliciesData;
